feat: Add session timeout background service
Add a new background service called SessionTimeoutBackgroundService to handle checking for false logged-in users. This service runs every 5 minutes and checks if the last login time of a user exceeds the session timeout. If it does, the user is logged out.

Also, update the infrastructure service registration to include the SessionTimeoutBackgroundService as a hosted service.