feat: Add endpoint to get devices status
Add a new endpoint to the DeviceController that allows retrieving the status of multiple devices. The endpoint accepts an array of device IDs, as well as optional pagination parameters for page number and page size. The endpoint returns a list of HardwareResponseDTO objects containing the device ID, status, and brightness (if applicable).

This commit adds the following changes:
- Adds a new action method GetDevicesStatus to the DeviceController
- Implements the corresponding query handler GetHardwareResponseQueryHandler
- Adds the HardwareResponseDTO class to the DeviceService.Application/DTOs namespace

Refs: #123