refactor(automation-service): Update Routine and RoutineExecutionLog entities
- Update Routine entity properties:
  - Change DeviceId property to nullable Guid
  - Change Status property to nullable DeviceStatus
  - Add ActionTarget property

- Update RoutineExecutionLog entity properties:
  - Add DeviceId property
  - Add ActionTarget property
  - Add ActionStatus property
  - Add ActionType property
  - Update ExecutedAt property to use UTC time

This commit refactors the Routine and RoutineExecutionLog entities in the automation-service. It updates the properties of the Routine entity to make DeviceId and Status nullable, adds the Brightness and ActionTarget properties. For the RoutineExecutionLog entity, it adds the DeviceId, ActionTarget, ActionStatus, and ActionType properties. Additionally, it updates the ExecutedAt property to use UTC time.